26-API: Re-organize testing files / folders
- move in-memory setup stuff into a separate folder
- move integration setup stuff into a separate folder
